Fix all warnings when building in arduino IDE

+++ repl.c
@@ -10,7 +10,7 @@
 extern Word* compiling;
 void repl() {
   while(true) {
-    printf_P(PSTR("[%u%s "), STACKP, compiling ? ">>" : "]");
+    printf_P(PSTR("[%lu%s "), (long unsigned int)STACKP, compiling ? ">>" : "]");
 
     if (tty_peek() == EOF) break;
     if (!lex_input()) continue;
